Try to write from Room and read from Scala

+++ wire-android-sync-engine/zmessaging/src/test/java/com/waz/db/ZMessagingDBTest.java
@@ -4,9 +4,14 @@ import android.content.Context;
 
 import androidx.annotation.CallSuper;
 import androidx.annotation.NonNull;
+import androidx.sqlite.db.SupportSQLiteDatabase;
 
+import com.waz.content.ZmsDatabase;
+import com.waz.model.UserData;
+import com.waz.model.UserId;
 import com.waz.service.tracking.DummyTrackingService;
 import com.waz.zclient.storage.db.UserDatabase;
+import com.waz.zclient.storage.db.users.model.UserEntity;
 import com.waz.zclient.storage.di.StorageModule;
 
 import org.junit.After;
@@ -16,30 +21,52 @@ import org.robolectric.Robolectric;
 import org.robolectric.RobolectricTestRunner;
 import org.robolectric.annotation.Config;
 
+import scala.Option;
+
 @Config(manifest = Config.NONE)
 @RunWith(RobolectricTestRunner.class)
 public class ZMessagingDBTest {
 
-    private static String DB_NAME = "test_db";
+    private static String DB_NAME = "14897932743491238932723984723142341234";
 
     private ZMessagingDB zMessagingDB;
+    private ZmsDatabase zmsDatabase;
     protected Context context;
 
     @CallSuper
     @Before
     public void beforeTest() {
         context = Robolectric.application;
-        zMessagingDB = new ZMessagingDB(context, DB_NAME, new DummyTrackingService());
     }
 
     @CallSuper
     @After
     public void afterTest() {
-        zMessagingDB.close();
-        context.getDatabasePath(zMessagingDB.getDatabaseName()).delete();
+        if (zMessagingDB != null) {
+            zMessagingDB.close();
+            context.getDatabasePath(zMessagingDB.getDatabaseName()).delete();
+        }
+        if (zmsDatabase != null) {
+            zmsDatabase.dbHelper().close();
+            context.getDatabasePath(zmsDatabase.dbHelper().getDatabaseName()).delete();
+        }
     }
 
-    protected void closeLegacyDatabase() {
+    private SupportSQLiteDatabase getZMessagingSQLiteDatabase() {
+        if (zMessagingDB == null) {
+            zMessagingDB = new ZMessagingDB(context, DB_NAME, new DummyTrackingService());
+        }
+        return zMessagingDB.getWritableDatabase();
+    }
+
+    private SupportSQLiteDatabase getZmsSQLiteDatabase() {
+        if (zmsDatabase == null) {
+            zmsDatabase = new ZmsDatabase(UserId.apply(DB_NAME), context, new DummyTrackingService());
+        }
+        return zmsDatabase.dbHelper().getWritableDatabase();
+    }
+
+    protected void closeZMessagingDatabase() {
         zMessagingDB.close();
     }
 
@@ -48,6 +75,11 @@ public class ZMessagingDBTest {
     }
 
     protected void insertLegacyUserData(@NonNull String userId, @NonNull String name) {
-        ZMessagingOperator.insertUserData(zMessagingDB, userId, name);
+        LegacyDaoUtil.insertUserData(getZMessagingSQLiteDatabase(), userId, name);
+    }
+
+    protected boolean verifyLegacyUserDataRead(final UserEntity userEntity) {
+        Option<UserData> userData = LegacyDaoUtil.getUserData(getZmsSQLiteDatabase(), userEntity.getId());
+        return LegacyDaoUtil.verifyUserData(userData, userEntity);
     }
 }
